requirements.txt

Django==5.2.12
djangorestframework==3.16.1
django-cors-headers==4.9.0
djangorestframework_simplejwt==5.5.1
channels==4.3.2
channels_redis==4.3.0
daphne==4.2.1
redis==7.3.0
python-decouple==3.8
whitenoise==6.12.0
